made surface stroke width double

+++ screwbox-core/src/main/java/dev/screwbox/core/environment/fluids/FluidRenderComponent.java
@@ -38,7 +38,7 @@ public class FluidRenderComponent implements Component {
      *
      * @since 3.20.0
      */
-    public int surfaceStrokeWidth = 2;
+    public double surfaceStrokeWidth = 2;
 
     /**
      * Draw order used for rendering.
